+++ cuckoo-service-api/cuckoo-system-service-api/src/main/java/org/github/zhengchalei/service/system/SysRoleService.java
@@ -1,8 +1,8 @@
 package org.github.zhengchalei.service.system;
 
-import org.github.zhengchalei.request.SysRoleSaveRequest;
-import org.github.zhengchalei.request.SysRoleUpdateRequest;
-import org.github.zhengchalei.response.SysRoleResponse;
+import org.github.zhengchalei.entity.system.request.SysRoleSaveRequest;
+import org.github.zhengchalei.entity.system.request.SysRoleUpdateRequest;
+import org.github.zhengchalei.entity.system.response.SysRoleResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
@@ -11,15 +11,15 @@ import java.util.List;
 public interface SysRoleService {
     Page<SysRoleResponse> findSysRolePage(Pageable pageable, String name);
 
-    SysRoleResponse findSysRoleById(Integer id);
+    SysRoleResponse findSysRoleById(Long id);
 
     SysRoleResponse saveSysRole(SysRoleSaveRequest sysRoleSaveRequest);
 
-    SysRoleResponse updateSysRoleById(Integer id, SysRoleUpdateRequest sysRoleUpdateRequest);
+    SysRoleResponse updateSysRoleById(Long id, SysRoleUpdateRequest sysRoleUpdateRequest);
 
-    SysRoleResponse partialSysRoleById(Integer id, SysRoleUpdateRequest sysRoleUpdateRequest);
+    SysRoleResponse partialSysRoleById(Long id, SysRoleUpdateRequest sysRoleUpdateRequest);
 
-    void deleteSysRoleById(Integer id);
+    void deleteSysRoleById(Long id);
 
     List<SysRoleResponse> findSysRoleList(String name);
